requirements.txt
# Core Dependencies
requests
python-dotenv
ipykernel

# Web App Framework
streamlit

# Data Processing
pandas
numpy

# Machine Learning
scikit-learn

# Visualization
matplotlib
seaborn

# Optional: Advanced ML (uncomment if needed)
# xgboost==2.0.3
# lightgbm==4.1.0

